requirements.txt
python-telegram-bot==21.4
fastapi==0.112.0
uvicorn==0.30.0
